requirements.txt
numpy>=1.26
pandas>=2.2
scipy>=1.11
PyYAML>=6.0
torch>=2.0
torchdiffeq>=0.2
fastapi>=0.100
uvicorn>=0.20
streamlit>=1.25
pytest>=7.4
black>=23.0
flake8>=6.0
docker
